Fix Long type errors and change buildAll .jar names

+++ meal/src/main/java/pl/auctane/meal/controllers/MealIngredientController.java
@@ -42,7 +42,7 @@ public class MealIngredientController {
     }
 
     @GetMapping(value = "/get/{mealId}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> getMealIngredient(@PathVariable("mealId") int mealId) {
+    public ResponseEntity<?> getMealIngredient(@PathVariable("mealId") Long mealId) {
         ObjectNode JSON = objectMapper.createObjectNode();
 
         //check if meal exist
